+++ .env.example
@@ -1,0 +1,3 @@
+DB_URL="jdbc:mysql://localhost:3312/fast_calculate_game"
+DB_USER="root"
+DB_PASSWORD="root"
