+++ BTL_LTM_SERVER/.idea/misc.xml
@@ -9,7 +9,7 @@
       </list>
     </option>
   </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_17" default="true" project-jdk-name="17" project-jdk-type="JavaSDK">
-    <output url="file://$PROJECT_DIR$/../../../../../../Cánhân/School/Năm 4 PTIT/LTM/LTM_Nhom15/BTL_LTM_SERVER/out" />
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_22_PREVIEW" project-jdk-name="17" project-jdk-type="JavaSDK">
+    <output url="file://$PROJECT_DIR$/.." />
   </component>
 </project>
