--- BTL_LTM_SERVER/cp.txt
@@ -1,1 +1,0 @@
-/Users/yourboyfriend/.m2/repository/com/mysql/mysql-connector-j/8.0.33/mysql-connector-j-8.0.33.jar:/Users/yourboyfriend/.m2/repository/com/google/protobuf/protobuf-java/3.21.9/protobuf-java-3.21.9.jar
