+++ BTL_LTM_SERVER/pom.xml
@@ -9,14 +9,24 @@
     <artifactId>BTL_LTM_SERVER</artifactId>
     <version>1.0-SNAPSHOT</version>
 
+
+
     <build>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
+                    <source>9</source>
+                    <target>9</target>
+                </configuration>
+            </plugin>
+            <plugin>
+                <groupId>org.codehaus.mojo</groupId>
+                <artifactId>exec-maven-plugin</artifactId>
+                <version>3.5.1</version>
+                <configuration>
+                    <mainClass>server.Main</mainClass>
                 </configuration>
             </plugin>
         </plugins>
@@ -26,7 +36,7 @@
         <dependency>
             <groupId>com.mysql</groupId>
             <artifactId>mysql-connector-j</artifactId>
-            <version>8.0.33</version>
+            <version>8.3.0</version>
         </dependency>
     </dependencies>
     <properties>
