+++ BTL_LTM_SERVER/src/main/java/server/Server.java
@@ -1,4 +1,4 @@
-package org.example;
+package server;
 import java.io.*;
 import java.net.Socket;
 import java.net.ServerSocket;
